[tested] upgrade dependencies.
- Tested on upgraded `cmake` version 4.0.3
- Upgrade dependencies:
	- `libigl`: v2.6.0
	- `geogram`: v1.9.6
	- `spdlog`: v1.15.3
	- `fmt`: 11.2.0 (for `spdlog`)
	- `CLI11`: v2.5.0
	- `polyscope`: v2.4.0
- Fixed many spdlog and fmt related errors, add a new file `include/matfp/LoggerFormatter.h`
- add CMAKE_OSX_ARCHITECTURES to x86_64

+++ src/matfp/InscribedSpheres.cpp
@@ -285,7 +285,8 @@ void collect_and_update_DT_feature_tets(
       TET_TYPE tet_type = get_nonfeature_tet_type(tet_vs_tags, tet_vs_normals,
                                                   tet_vs_bucket, is_debug);
       if (is_debug) {
-        logger().error("cell_tag {} tet_type: {}", cell_tag, tet_type);
+        logger().error("cell_tag {} tet_type: {}", cell_tag,
+                       static_cast<int>(tet_type));
       }
       // 4:0 is sliver, we do not know how to handle
       if (tet_type == TET_TYPE::T4_0) {
